moved get_flange_limits to accessories
